api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-bigquery==3.25.0
pydantic==2.7.4
